backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
supabase==2.7.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Payments
stripe==8.2.0

# Email
sendgrid==6.11.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
isort==5.13.2
httpx==0.26.0
